Update apache php

+++ apache-php/entrypoint.sh
@@ -4,11 +4,11 @@ set -e
 
 if [ "$1" = "apache2-foreground" ]; then
 
-    if [ "${APACHE_ENABLE_HTTP_TRAFFIC}" = "false" ]; then
+    if [ "${APACHE_ENABLE_HTTP_TRAFFIC}" = "false" ] && [ -f "/etc/apache2/sites-available/000-default.conf" ]; then
         rm -f /etc/apache2/sites-available/000-default.conf
     fi
 
-    if [ "${APACHE_ENABLE_HTTPS_TRAFFIC}" = "false" ]; then
+    if [ "${APACHE_ENABLE_HTTPS_TRAFFIC}" = "false" ] && [ -f "/etc/apache2/sites-available/default-ssl.conf" ]; then
         rm -f /etc/apache2/sites-available/default-ssl.conf
     fi
